fabrique/pyproject.toml
[tool.poetry]
name = "fabrique"
version = "0.1.0"
description = "Test task"
authors = ["vadikam100500 <vadik.msk.100500@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^0.19.2"
Django = "^4.0.2"
djangorestframework = "^3.13.1"
drf-yasg = "^1.20.0"
django-extensions = "^3.1.5"
psycopg2-binary = "^2.9.3"
djoser = "^2.1.0"
django-filter = "^21.1"
requests = "^2.27.1"
gunicorn = "^20.1.0"
celery = "^5.2.3"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
flake8-broken-line = "^0.4.0"
flake8-return = "^1.1.3"
flake8-isort = "^4.1.1"
pep8-naming = "^0.12.1"
mypy = "^0.931"
isort = "^5.10.1"
pylint = "^2.12.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
